refactor: folder structs and handle login
