+++ src/backend/src/adapters.rs
