+++ src/backend/src/adapters/user_repository.rs
